+++ src/app/pages/aprobacion-ordenespago/components/budget-approbation/budget-approbation.component.html
@@ -1,1 +1,36 @@
-<p>budget-approbation works!</p>
+<div class="container-fluid">
+    <div class="card">
+        <div class="card-body">
+            <div class="row">
+                <div class="col py-2">
+                    <h5 class="card-title text-center text-primary">APROBACIÓN PRESUPUESTAL DE ÓRDENES DE PAGO</h5>
+                </div>
+            </div>
+            <div class="card">
+                <div class="card-body">
+                    <div class="card-title col-md-4 bg-primary text-light text-center py-1 px-1 rounded">
+                        Aprobación presupuestal
+                    </div>
+                    <div class="table">
+                        <ngx-table [titles]="titles" [attributes]="attributes" [list]="datosCreacion" [aprobacionContable]="false" [aprobacionPresupuestal]="false" [seleccion]="false" > </ngx-table>
+                    </div>
+                    <div class="row">
+                        <div class="col">
+                            <button type="button" class="btn btn-outline-dark px-3 py-2 my-2" (click)="rechazarAprobacion()">
+                                Rechazar
+                            </button>
+                        </div>
+                        <div class="col">
+                            <button type="button" class="btn btn-success px-3 py-2 my-2 text-light" (click)="aceptarAprobacion()">
+                                Aprobar
+                            </button>
+                        </div>
+                    </div> 
+                </div>
+            </div>
+            
+        </div>
+    </div>
+</div>
+
+
