+++ src/app/pages/aprobaciones/aprobaciones.module.ts
@@ -26,9 +26,6 @@ import { MatCardModule,
   MatTableModule } from '@angular/material';
 import { LayoutComponent } from './components/layout/layout.component';
 import { RechazarComponent } from './components/rechazar/rechazar.component';
-import { BudgetApprobationComponent } from './components/budget-approbation/budget-approbation.component';
-import { AccountantApprobationComponent } from './components/accountant-approbation/accountant-approbation.component';
-
 
 @NgModule({
   declarations: [
@@ -37,7 +34,7 @@ import { AccountantApprobationComponent } from './components/accountant-approbat
     ListApprobationsComponent,
     RelationApprobationComponent,
     OrderApprobationComponent,
-    FormComponent, LayoutComponent, RechazarComponent, BudgetApprobationComponent, AccountantApprobationComponent],
+    FormComponent, LayoutComponent, RechazarComponent],
   imports: [
     CommonModule,
     SharedModule,
