--- src/app/pages/aprobaciones/components/accountant-approbation/accountant-approbation.component.html
@@ -1,1 +1,0 @@
-<p>accountant-approbation works!</p>
