--- src/app/pages/aprobaciones/components/accountant-approbation/accountant-approbation.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'ngx-accountant-approbation',
-  templateUrl: './accountant-approbation.component.html',
-  styleUrls: ['./accountant-approbation.component.scss']
-})
-export class AccountantApprobationComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
