--- src/app/pages/aprobaciones/components/budget-approbation/budget-approbation.component.html
@@ -1,31 +1,0 @@
-<div class="container-fluid">
-    <div class="card">
-        <div class="card-body">
-            <div class="card-title col-md-4 bg-primary text-light text-center py-1 px-1 rounded">
-                Aprobación presupuestal
-            </div>
-            <div class="row">
-                <p> Seleccione las órdenes que desee aprobar: </p>
-            </div>
-            <div class="table">
-                <ngx-table [titles]="titles" [attributes]="attributes" [list]="datosCreacion" [seleccion]="true" [aprobacionContable]="true" [aprobacionPresupuestal]="true"> </ngx-table>
-            </div>
-            <div class="row">
-                <div class="col">
-                    <label> Aprobación masiva </label>
-                </div>
-                <div class="col">
-                    <button type="button" class="btn btn-outline-dark px-3 py-2 my-2" (click)="recha()">
-                        Aprobación contable
-                    </button>
-                </div>
-                <div class="col">
-                    <button type="button" class="btn btn-success px-3 py-2 my-2 text-light" (click)="aprobar()">
-                        Aprobación presupuestal
-                    </button>
-                </div>
-            </div> 
-        </div>
-    </div>
-</div>
-
