+++ src/app/pages/aprobaciones/interfaces/interfaces.ts
@@ -238,51 +238,4 @@ export const DATOS_ORDENPAGO: any = [
         nombre: 'Nancy'
      },
 ];
-export const DATOS_ORDEN_DETALLE: any = [
-    {
-        codigoCuentaContable: '123',
-        vigencia: '2019',
-        entidad  : 'Universidad',
-        unidad: 'Tesoreria',
-        disponibilidad: 'Si',
-        registro: '123',
-        valor: '139473'
-    },
-    {
-        codigoCuentaContable: '2743',
-        vigencia: '2020',
-        entidad  : 'Otra',
-        unidad: 'Contable',
-        disponibilidad: 'Si',
-        registro: '5534',
-        valor: '829374'
-     },
-     {
-        codigoCuentaContable: '7656',
-        vigencia: '2018',
-        entidad  : 'Hotel',
-        unidad: 'Extension',
-        disponibilidad: 'No',
-        registro: '534',
-        valor: '131231'
-     },
-     {
-        codigoCuentaContable: '123',
-        vigencia: '2020',
-        entidad  : 'Colegio',
-        unidad: 'Tesoreria',
-        disponibilidad: 'Si',
-        registro: '4234',
-        valor: '434234'
-     },
-     {
-        codigoCuentaContable: '123',
-        vigencia: '2019',
-        entidad  : 'Distrital',
-        unidad: 'Otro',
-        disponibilidad: 'Si',
-        registro: '4534',
-        valor: '123123'
-     },
-];
